First pixel drawn on screen

+++ emu_6502/src/machine/register.rs
@@ -1,15 +1,15 @@
 #[derive(Clone, Copy)]
 pub struct Register {
     // TODO i16
-    pub pc: i8,
+    pub pc: i32,
     // accumulator
-    pub ac: i8,
-    x: i8,
-    y: i8,
+    pub ac: i32,
+    x: i32,
+    y: i32,
     // Status register
-    sr: i8,
+    sr: i32,
     //  Stack pointer
-    sp: i8
+    sp: i32
 }
 
 impl Register {
